Fixed lazy/eager loading of filteroutcomes
